requirements.txt
torch==2.9.0
torchvision==0.24.0
numpy<2.0
scipy
pandas
tqdm
typing_extensions
monai
fvcore
timm
einops
open_clip_torch
opencv-python
scikit-image
scikit-learn
albumentations
ogb>=1.2.6
SimpleITK>=2.2.1
connected-components-3d
nibabel
transformers
ftfy
regex
sentencepiece
matplotlib
openpyxl
easydict
huggingface_hub